Move `get_sparse_forces` to a more general location
