+++ python-libraries/nanover-openmm/src/nanover/openmm/imd.py
@@ -64,6 +64,7 @@ from nanover.imd.imd_force import calculate_imd_force
 from nanover.imd import ImdStateWrapper
 from nanover.trajectory.frame_publisher import FramePublisher
 from nanover.imd.particle_interaction import ParticleInteraction
+from nanover.imd.imd_force import get_sparse_forces
 from .converter import openmm_to_frame_data
 from nanover.trajectory.frame_data import Array2Dfloat, FrameData
 
@@ -372,24 +373,3 @@ def get_imd_forces_from_system(system: Simulation) -> List[CustomExternalForce]:
         and force.getEnergyFunction() == IMD_FORCE_EXPRESSION
         and force.getNumParticles() == system_num_particles
     ]
-
-
-def get_sparse_forces(user_forces: npt.NDArray) -> Tuple[npt.NDArray, npt.NDArray]:
-    """
-    Takes in an array of user forces acting on the system containing N particles
-    and outputs two arrays that describe these user forces in a sparse form:
-
-    - The first contains the indices of the particles for which the user forces
-      are non-zero
-    - The second contains the non-zero user forces associated with each index
-
-    :param user_forces: Array of user forces with dimensions (N, 3)
-    :return: Array of particle indices, Array of corresponding user forces
-    """
-    sparse_indices = np.unique(np.nonzero(user_forces)[0])
-    sparse_forces = np.zeros((sparse_indices.shape[0], 3))
-
-    for index in range(sparse_indices.shape[0]):
-        sparse_forces[index, :] = user_forces[sparse_indices[index]]
-
-    return sparse_indices, sparse_forces
